Move data converter to widgets module.
